+++ src/cm/mod.rs
@@ -7,4 +7,5 @@
 
 pub(crate) mod block;
 pub(crate) mod inline;
+pub(crate) mod math;
 pub(crate) mod refs;